notesAnonymizer now takes list && changed name replacement to rand int
the function can now take a list of assessments directly from the graphql request and return it anonymized. the name replacement has also recieved an improvement as it now sets a random int betw. 1-100 to users and keeps the same numbers for every CODE user (so in all of someones assessments, if two people who took part in the same collective ass. submit their LP data their names will be different to one another but consistent in all of their OWN assessment notes.)